login auth working
