+++ client/src/components/Nav/index.css
@@ -1,0 +1,4 @@
+
+.small-font {
+ font-size: 12px;
+}
