+++ client/src/components/signup-modal/modal.css
@@ -1,0 +1,87 @@
+.modal {
+ position: fixed;
+ top: 0;
+ left: 0;
+ width: 100%;
+ height: 100%;
+ background-color: rgba(0, 0, 0, 0.5);
+ display: none;
+ justify-content: center;
+ align-items: center;
+}
+
+.modal.open {
+ display: flex;
+}
+
+.modal-content {
+ background-color: white;
+ padding: 20px;
+ border-radius: 5px;
+ max-width: 400px;
+ width: 100%;
+}
+
+.close-button {
+ position: absolute;
+ top: 10px;
+ right: 10px;
+ cursor: pointer;
+}
+
+form {
+ display: flex;
+ flex-direction: column;
+}
+
+label {
+ margin-bottom: 5px;
+}
+
+input {
+ padding: 10px;
+ margin-bottom: 10px;
+ border-radius: 5px;
+ border: 1px solid #ccc;
+}
+
+button {
+ padding: 10px 20px;
+ background-color: #007bff;
+ color: white;
+ border: none;
+ border-radius: 5px;
+ cursor: pointer;
+}
+
+.gold-text {
+    color: #a68e45
+}
+
+.light-text {
+    color: #a28e7b;
+}
+
+.dark-text {
+    color: #231a11;
+}
+
+
+.gold-background {
+    background-color: #a68e45;
+}
+
+.light-background {
+    background-color: #a28e7b;
+}
+
+.dark-background {
+    background-color: #231a11;
+}
+
+.success-message {
+ color: white;
+ /* background-color: green;  */
+ padding: 10px;
+ border-radius: 5px;
+}
