+++ client/src/main.jsx
@@ -4,7 +4,6 @@ import './styles.css'
 
 import App from './App.jsx'
 import Home from './pages/Home';
-import Login from './pages/Login';
 import BookingHistory from './pages/BookingHistory'
 import Booking from './pages/Booking'
 import Confirmation from './pages/Confirmation'
@@ -19,9 +18,6 @@ const router = createBrowserRouter([
       {
         index: true,
         element: <Home />
-      }, {
-        path: '/login',
-        element: <Login />
       }, 
        {
         path: '/BookingHistory',
